fix: use cupertino
```
Expected to find fonts for (MaterialIcons, packages/cupertino_icons/CupertinoIcons), but found (MaterialIcons). This usually means you are referring to font
families in an IconData class but not including them in the assets section of your pubspec.yaml, are missing the package that would include them, or are
missing "uses-material-design: true".
```